main: remove use of DONT_CATCH
With removing this macro, global_EC doesn't make much sense anymore since it
was mainly used by exception handler to signal that there has been an error and
exit with this. I'm going to use actual exceptions instead in order to preserve
the backtrace in case error is occured. This program is eons from being in
production, so it's fine like that.