gl_texturecache: fix usage of iterators
This commit is a result of debugging the Windows debug runtime assertion
failure after pressing F6 in viewer.exe. Turned out the removing of the
item while iterating through the collection contained a mistake that
caused invalid memory access. While I was at it, I simplified a couple
of other loops to use for-each instead of plain iterators.