gl_cityvew: fix uint8_t underflow
This issue was found by Windows debug runtime raising an error about
invalid vector access. The size of vector in question was 1, the
frame_num was 0, hence `frame_num - 1` was `255`.
This is not the proper fix yet, however. So far I just made it work
without invalid memory access, but this file definitely needs some
refactoring.