util: added bitwise arithmetic header-only lib
When proper support for modules is introduce, it should be converted to
a module, because right now every change in it requires a recompilation
of many source files. Before that, I might consider having a PCH.
Additionally, a GoogleTest file for this module has been added to test
the functionality.