viewer: Fix crash on exit
The crash was introduced during an attempt to reduce the amount of
global variables and removing of atexit from viewer, which ended up with
creation of OpenGTAViewer struct that contains everything needed to
start the program.
Sadly, to this day it uses some of the singletons that are destroyed at
the very end. However, PHYSFS_deinit was called before that, which
caused an application to crash on ActiveData destructor. Hence, I'll
change viewer to leak intentionally in order to fix the crash until I
get rid of ActiveData singleton.