Remove conan dependency
Conan is a nice package manager. However, as can be observed with these
changes, it requires a developer to change the invocations of
`find_package` and `target_link_packages` even for `CMakeToolchain`
generator.
This, imo, contradicts the value of `CMakeToolchain` itself, as many
(myself included) used this generator to have flexibility between the
scenarios where dependencies are handled with Conan and with system
package manager.

This commit removes `conanfile.py` and the corresponding
`CMakeUserPresets.json` and rolls back the dependencies in
`CMakeLists.txt`. From now on, the dependencies will be handled via
system package manager or docker. Or Nix.