Refactor PhysFSContext constructor and add new file methods
This commit refactors the `PhysFSContext` constructor in the `file-manager.cpp` file to accept two new boolean parameters: `mount_base_dir` and `mount_gtadata`. It also adds three new methods to the `PhysFSFile` class: `length()`, `tell()`, and `eof()`. These methods provide functionality to retrieve the length of a file, the current position in the file, and whether the end of the file has been reached, respectively.

The changes in this commit improve the flexibility and functionality of the file management system in the codebase.